+++ src/features/profile/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as ProfilePage } from './profilePreview/ProfilePreview.tsx';
