+++ src/features/profile/ui/profilePreview/ProfilePreview.tsx
@@ -1,0 +1,3 @@
+const ProfilePreview = () => {};
+
+export default ProfilePreview;
