+++ src/features/project/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectCard } from './projectCard/ProjectCard.tsx';
