+++ src/features/signup/index.ts
@@ -1,1 +1,2 @@
-export { useSignupFunnel } from './model/useSignupFunnel';
+export * from './ui';
+export * from './model';
