+++ src/pages/index.ts
@@ -1,1 +1,2 @@
 export { default as SignUpPage } from './signupPage/SignUpPage.tsx';
+export { default as MainPage } from './mainPage/MainPage.tsx';
