+++ src/pages/mainPage/index.ts
@@ -1,0 +1,1 @@
+export { default as MainPage } from './MainPage.tsx';
