+++ src/pages/mainPage/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as MainPageHeader } from './MainPageHeader.tsx';
