+++ src/pages/signupPage/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as SignUpPageHeader } from './signUpPageHeader/SignUpPageHeader.tsx';
