+++ src/shared/ui/button/index.ts
@@ -1,3 +1,4 @@
 // Button components
 export { default as IconButton } from './iconButton/IconButton';
 export { default as TextButton } from './textButton/TextButton';
+export { default as FloatButton } from './floatButton/FloatButton';
