+++ src/shared/ui/notificationModal/content/index.ts
@@ -1,0 +1,1 @@
+export { default as ApplicationNotification } from './applicationNotification/ApplicationNotification.tsx';
